pyproject.toml
[project]
name = "openclaw-alpaca-trader"
version = "0.1.0"
description = "Autonomous AI trading agent using OpenClaw + Alpaca Paper Trading"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.35.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "ta>=0.11.0",
    "vectorbt>=0.26.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
